An updated version of Alek's automated function
I added a lot of annotation, as well as automated checks to make sure that the user was providing the correct input, and little print messages regarding each of the output.  It needs to be checked for bugs and edited - the little automated messages are currently really hard to read.  Maybe I can add numbers, spaces, and bulleting?  Or perhaps stick most of the intermediate files into a separate subdirectory?